Complete the slient install for subiquity
Using the casper initrd hook to enable/disable subiquity, which in headless_install or factory_install mode.
Added a curtin-cfg to edit the curtin config file which generated
by subiquity. The curtin-wrapper will copy to /usr/local/bin/ which will
be called first. curtin-wrapper calls curtin-cfg and recovery.bin and
then the real curtin to complete the silent install.